+++ gauguin-core/src/test/kotlin/org/piepmeyer/gauguin/grid/GridTest.kt
@@ -51,11 +51,24 @@ class GridTest : FunSpec({
         grid.getValidCellAt(1, 0).duplicatedInRowOrColumn shouldBe false
         grid.getValidCellAt(1, 1).duplicatedInRowOrColumn shouldBe false
     }
+
+    test("correct and incorrect values in one row should lead to exactly 1 mistake") {
+        val grid = smallGrid()
+
+        grid.getValidCellAt(0, 0).userValue = 2
+        grid.getValidCellAt(0, 1).userValue = 2
+
+        println(grid)
+
+        grid.numberOfMistakes() shouldBe 1
+    }
 })
 
 private fun smallGrid(): Grid {
     return GridBuilder(2)
         .addCage(2, GridCageAction.ACTION_MULTIPLY, GridCageType.ANGLE_RIGHT_BOTTOM, 0)
         .addSingleCage(2, 3)
+        .addValueRow(2, 1)
+        .addValueRow(1, 2)
         .createGrid()
 }